# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dreamers Solutions ERP APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "operationId": "64e7f64d53050c5582e7f6e59e2e4285",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a category",
                "operationId": "8cdcc1c8ebab5e0be2ba6d4ac19969b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/active": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "operationId": "fdf709f098475eced3fac2590a393ddb",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category by it's id",
                "operationId": "f8a52cfc06ed1b07aaac7e2171c60d56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the category",
                        "required": true,
                        "allowEmptyValue": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid category id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update a category",
                "operationId": "7e08582077836097a42794366a76ddd6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the category",
                        "required": true,
                        "allowEmptyValue": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category",
                "operationId": "f846e11051f688ef53ade713d30cb068",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the category",
                        "required": true,
                        "allowEmptyValue": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid category id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/{id}/active": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Activate a deleted category",
                "operationId": "20fcc671913fb75cbd023b76113cb8b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the category",
                        "required": true,
                        "allowEmptyValue": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid category id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/companies/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get company details",
                "operationId": "fbaf0a639542a6c9b1bb7bf06209944a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategoryRequest": {
                "title": "CategoryRequest",
                "description": "Category request body",
                "required": [
                    "categoryName",
                    "active"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the category",
                        "type": "integer"
                    },
                    "categoryName": {
                        "description": "Name of the category",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the category",
                        "type": "string"
                    },
                    "parentId": {
                        "description": "Parent category id of the category",
                        "type": "integer"
                    },
                    "active": {
                        "description": "Active status of the category",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CategoryResponse": {
                "title": "CategoryResponse",
                "description": "Category response model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the category",
                        "type": "integer"
                    },
                    "categoryName": {
                        "description": "Name of the category",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the category",
                        "type": "string"
                    },
                    "parent": {
                        "$ref": "#/components/schemas/CategoryRequest"
                    },
                    "children": {
                        "description": "Active status of the category",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryRequest"
                        }
                    },
                    "active": {
                        "description": "Active status of the category",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CompanyResponse": {
                "title": "CompanyResponse",
                "description": "Company details",
                "properties": {
                    "id": {
                        "description": "Company id",
                        "type": "integer"
                    },
                    "company_name": {
                        "description": "Company name",
                        "type": "string"
                    },
                    "company_address1": {
                        "description": "Address line 1",
                        "type": "string"
                    },
                    "company_address2": {
                        "description": "Address line 2",
                        "type": "string"
                    },
                    "company_address3": {
                        "description": "Address line 3",
                        "type": "string"
                    },
                    "company_tel1": {
                        "description": "Telephone 1",
                        "type": "string"
                    },
                    "company_tel2": {
                        "description": "Telephone 2",
                        "type": "string"
                    },
                    "company_fax": {
                        "description": "Fax number",
                        "type": "string"
                    },
                    "company_email": {
                        "description": "Company email",
                        "type": "string"
                    },
                    "company_web": {
                        "description": "Company website",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token obtained from the login endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}